wms/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==2.8
datadog-api-client==2.21.0
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.1
django-cors-headers==4.3.1
django-mpesa==2.0.14
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
fabric==3.2.2
Fabric3==1.14.post1
filelock==3.13.1
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==3.0.1
html5lib==1.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
invoke==2.2.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
MarkupSafe==2.1.3
mistune==3.0.2
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
paramiko==2.12.0
pep8==1.7.1
pipenv==2023.11.15
platformdirs==4.1.0
pycodestyle==2.8.0
pycparser==2.21
PyGObject==3.42.1
PyJWT==2.8.0
PyNaCl==1.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.4
PyYAML==5.4.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.16.2
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
soupsieve==2.3.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
virtualenv==20.25.0
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==1.0.0
